fix: encode clientbound packet ids as VarInt instead of u8
In MC 26.1 (protocol 775), seven new clientbound play packets were
inserted (game_rule_values, low_disk_space_warning, mount_screen_open,
attack, change_game_mode, set_game_rule, spectate_entity), pushing
several existing packets past id 127. minecraft:transfer in particular
went from 127 → 129. RawPacket previously stored packet_id as a single
raw u8, which is wire-compatible with a VarInt only for ids 0..=127.
For id 129 the server wrote 0x81, the client read 0x81 (continuation
bit set) plus the next payload byte as the rest of the varint, and
disconnected with "unknown packet id 3201" — the bytes 0x81 0x19
decoded as VarInt(3201), where 0x19 was the host string length of the
Transfer payload.

RawPacket now stores [VarInt packet_id][payload] and exposes the
packet id as i32. The packet_reports macro generates i32 ids and
PacketRegistryDecodeError carries i32 instead of u8. PacketStream is
unchanged because bytes()/size() still return the full encoded body.

Confirmed reproducible against a 775 client and fixed end-to-end via
a TCP probe; play-state Transfer packet now goes out with 2-byte
VarInt 0x81 0x01 followed by the host string. Older versions (≤127)
remain a single byte and are bit-identical on the wire.

+++ crates/macros/src/packet_reports.rs
@@ -12,7 +12,7 @@ use syn::{Data, DeriveInput, Error, Fields, Ident, LitStr, Token, parse_macro_in
 /// Represents the "protocol_id" object within the JSON.
 #[derive(Debug, Deserialize)]
 struct PacketInfo {
-    protocol_id: u8,
+    protocol_id: i32,
 }
 
 /// Represents the mapping from packet_name to PacketInfo.
@@ -59,9 +59,9 @@ pub fn packet_report_derive(input: TokenStream) -> TokenStream {
             #[error("Decode error: Packet id is missing from the payload")]
             MissingPacketId,
             #[error("Decode error: Unknown version version={1} packet_id={0}")]
-            UnknownVersion(u8, i32),
+            UnknownVersion(i32, i32),
             #[error("Decode error: Packet not found version={0} state={1} packet_id={2}")]
-            NoCorrespondingPacket(i32, State, u8),
+            NoCorrespondingPacket(i32, State, i32),
             #[error("Failed to read packet")]
             Decode(#[from] BinaryReaderError),
         }
@@ -302,7 +302,7 @@ fn generate_encode_impl(
                 #enum_ident::#variant_ident(packet) => {
                     packet.encode(&mut packet_writer, protocol_version)?;
                     let packet_bytes = packet_writer.into_inner();
-                    let packet_id: u8 = match packets_version {
+                    let packet_id: i32 = match packets_version {
                         #(#report_arms)*
                         _ => return Err(PacketRegistryEncodeError::UnsupportedPacket(protocol_version, String::from(#packet_name))),
                     };
